add 'onReadyHandler' and 'onFeedPopulationHandler' config options
The onReadyHandler is called once after the tweets are initially loaded
and added to the DOM. Immediately after calling this event handler the
onFeedPopulationHandler is called with the invocations parameter set to
0.

The onFeedPopulationHandler is called each time new tweets were added
to the tweet feed - and thereby to the DOM. The supplied event handler
should have the following interface: function(invocations, options) {}
The invocations parameter contains the present # of calls to the
handler, starting at 0 for the first call. This event handler is called
either for populating the feed by paging or by autorefreshing.

issue #9