add config option "autorefresh.behavior"
This options controls the behavior of the plugin when new tweets are to
be inserted. Different approaches are eligible, depending on the
paging.mode and on the autorefresh.mode.

Accepted values for behavior are: "expand-list" | "keep-list-size" |
"move-to-inserted" | "keep-position"

The values "expand-list" and "keep-list-size" work in conjunction with
paging.mode "none" and "more".

If behavior equals "expand-list" (the default behavior for paging.mode
"none" and "more") new tweets will be inserted before existing ones,
resulting in a growing tweet feed.

If behavior equals "keep-list-size" then for each inserted tweet the
last displayed tweet will be removed, resulting in a constant size tweet
feed.

The values "move-to-inserted" and "keep-position" work in conjunction
with paging.mode "prev-next" and "endless-scroll".

If behavior equals "move-to-inserted" (the default behavior for
paging.mode "prev-next" and "endless-scroll", if autorefresh.mode equals
"trigger-insert") then after inserting new tweets to the tweet feed
these tweets will be brought into view. Keep in mind that if you use
this option if autorefresh.mode equals "auto-insert" your users might be
harshly interrupted while reading tweets.

If behavior equals "keep-position" (the default behavior for paging.mode
"prev-next" and "endless-scroll", if autorefresh.mode equals
"auto-insert") then new tweets will silently be inserted in the tweet
feed without changing the user's viewport.

issue #6
issue #7
issue #14